+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # CHANGELOG
+## [1.2.0] 2022-06-22
+* Compatible with CombatReady! 4.1.0
+* Added Video animation
+* Fixed bug that always show light meter
+
 ## [1.1.0] 2022-02-15
 * Updated module.json for v9
 * Compatible with CombatReady! 4.0.0